Conflicts:
	PyInstaller/configure.py
	PyInstaller/hooks/hook-raven.py

+++ bootloader/common/pyi_python.h
@@ -1,35 +1,22 @@
 /*
+ * ****************************************************************************
+ * Copyright (c) 2013, PyInstaller Development Team.
+ * Distributed under the terms of the GNU General Public License with exception
+ * for distributing bootloader.
+ *
+ * The full license is in the file COPYING.txt, distributed with this software.
+ * ****************************************************************************
+ */
+
+
+/*
  * Python.h replacements.
  *
  * We use dynamic loading -> one binary can be used with (nearly) any Python
  * version. This is the cruft necessary to do dynamic loading.
- *
- * Copyright (C) 2012, Martin Zibricky
- * Copyright (C) 2005-2011, Giovanni Bajo
- * Based on previous work under copyright (c) 2002 McMillan Enterprises, Inc.
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
- *
- * In addition to the permissions in the GNU General Public License, the
- * authors give you unlimited permission to link or embed the compiled
- * version of this file into combinations with other programs, and to
- * distribute those combinations without any restriction coming from the
- * use of this file. (The General Public License restrictions do apply in
- * other respects; for example, they cover modification of the file, and
- * distribution when not linked into a combine executable.)
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA
  */
+
+
 #ifndef PYI_PYTHON_H
 #define PYI_PYTHON_H
 
@@ -105,7 +92,7 @@ EXTDECLPROC(int, PyObject_SetAttrString, (PyObject *, char *, PyObject *));
 EXTDECLPROC(PyObject *, PyList_New, (int));
 EXTDECLPROC(int, PyList_Append, (PyObject *, PyObject *));
 EXTDECLPROC(PyObject *, Py_BuildValue, (char *, ...));
-EXTDECLPROC(PyObject *, PyString_FromStringAndSize, (const char *, int));
+EXTDECLPROC(PyObject *, PyString_FromStringAndSize, (const char *, size_t));
 EXTDECLPROC(PyObject *, PyFile_FromString, (char *, char *));
 EXTDECLPROC(char *, PyString_AsString, (PyObject *));
 EXTDECLPROC(PyObject *, PyObject_CallFunction, (PyObject *, char *, ...));
